+++ frontend/lost/src/actions/lost/index.js
@@ -1,3 +1,0 @@
-import * as lost from './lost'
-
-export default { ...lost }