+++ frontend/lost/src/actions/worker/index.js
@@ -1,3 +1,0 @@
-import * as workers from './workers'
-
-export default { ...workers }