+++ frontend/lost/src/containers/Users.jsx
@@ -1,7 +1,0 @@
-import React from 'react'
-
-const Users = () => {
-    return <div>Users</div>
-}
-
-export default Users