+++ frontend/lost/src/containers/Workers.jsx
@@ -1,7 +1,0 @@
-import React from 'react'
-
-const Workers = () => {
-    return <div>Workers</div>
-}
-
-export default Workers