+++ frontend/lost/src/reducers/index.js
@@ -1,20 +1,16 @@
 import { combineReducers } from 'redux'
 import { reducer as formReducer } from 'redux-form'
 import annoTask from './annoTask'
-import lost from './lost'
 import mia from './mia'
 import sia from './sia'
 import siaReview from './siaReview'
-import worker from './worker'
 
 const appReducer = combineReducers({
     annoTask,
     mia,
     sia,
-    worker,
     form: formReducer,
     siaReview,
-    lost,
 })
 
 const rootReducer = (state, action) => {